fec-1071 - fix custom scroller jump when focusing on number input. currently solved by disabling the focus scroll internal utility via internal option.
this requiered refactoring the menupage/navmenu to work with the mcustomscrollbar directive rather than directly apply the plugin. much better for code organisation but do notice that preformance wise the navmenu scroller (one for the whole menu) was better, I'm not sure on the actual impact. for my Chrome this is unnoticeable.
I kept the code & selectors used as comments in mCustomScrollbar.js in case you'll want to make a special end case and make the scroller destory and reinstiate based on visibilty